refactor: extract BatchRunner from Orchestrator
Move batch processing loop (runBatch, #runBatchInner) into a dedicated
BatchRunner class in batch-runner.ts. Orchestrator.runBatch() becomes a
thin delegation wrapper preserving the public API signature.

- Add SingleIssueRunner interface to decouple BatchRunner from Orchestrator
- Remove batch-only imports from orchestrator.ts (IssueSyncer, Prioritizer, Queue, etc.)
- Export BatchRunner and SingleIssueRunner from mod.ts
